third_party/WebKit/LayoutTests/inspector-protocol/debugger/debugger-setBreakpoint-before-enabling-expected.txt

setBreakpointByUrl error: undefined
setBreakpoint error: {
  "code": -32602,
  "message": "Some arguments of method 'Debugger.setBreakpoint' can't be processed",
  "data": [
    "Parameter 'location' with type 'Object' was not found."
  ]
}
